+++ src/modules/components/vote/proposalVoteDataListItem/proposalVoteDataListItemSkeleton/index.ts
@@ -1,4 +1,4 @@
 export {
-    ProposalVoteDataListItemSkeleton,
-    type IProposalVoteDataListItemSkeletonProps,
-} from './proposalVoteDataListItemSkeleton';
+    VoteProposalDataListItemSkeleton,
+    type IVoteProposalDataListItemSkeletonProps,
+} from './voteProposalDataListItemSkeleton';