+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
--   Implement `ProposalVoteDataListItem` with Structure and Skeleton module
+-   Implement `VoteProposalDataListItem` with Structure and Skeleton module
 
 ### Changed